+++ src/renderer/gallery.css
@@ -1,0 +1,134 @@
+/* Drop zone */
+#drop-zone {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: linear-gradient(135deg, #2a2a2a 0%, #1a1a1a 100%);
+  border: 2px dashed #444;
+  transition:
+    background-color 0.3s ease,
+    border-color 0.3s ease;
+}
+
+#drop-zone.drag-over {
+  background: linear-gradient(135deg, #3a3a3a 0%, #2a2a2a 100%);
+  border-color: #666;
+}
+
+.drop-zone-content {
+  text-align: center;
+  max-width: 400px;
+}
+
+.drop-icon {
+  font-size: 4rem;
+  margin-bottom: 1rem;
+  opacity: 0.7;
+}
+
+.drop-zone-content h2 {
+  font-size: 1.5rem;
+  margin-bottom: 0.5rem;
+  font-weight: 300;
+}
+
+.drop-zone-content p {
+  color: #ccc;
+  margin-bottom: 2rem;
+  font-size: 0.9rem;
+}
+
+#file-select-btn {
+  background: #007acc;
+  color: white;
+  border: none;
+  padding: 12px 24px;
+  border-radius: 6px;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+#file-select-btn:hover {
+  background: #005aa3;
+}
+
+/* Gallery grid */
+#gallery-grid {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  overflow-y: auto;
+  padding: 2px;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 2px;
+  align-items: start;
+  overflow-anchor: auto;
+  contain: layout style;
+}
+
+/* Gallery images */
+.gallery-image {
+  width: 100%;
+  height: auto;
+  display: block;
+  background: #1a1a1a;
+  cursor: pointer;
+  contain: layout style paint;
+  /* Disable dragging */
+  -webkit-user-drag: none;
+  -webkit-user-select: none;
+  user-select: none;
+  /* Performance optimizations */
+  -webkit-backface-visibility: hidden;
+  backface-visibility: hidden;
+}
+
+.image-error {
+  width: 100%;
+  min-height: 200px; /* Give error placeholders a minimum height */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: #2a1a1a;
+  color: #ff6b6b;
+  font-size: 0.9rem;
+  text-align: center;
+  padding: 1rem;
+  border-radius: 2px;
+  border: 1px solid #333;
+  cursor: pointer;
+  contain: layout style paint;
+}
+
+.image-error-icon {
+  font-size: 2rem;
+  margin-bottom: 0.5rem;
+}
+
+/* Scrollbar styling */
+::-webkit-scrollbar {
+  width: 8px;
+}
+
+::-webkit-scrollbar-track {
+  background: #1a1a1a;
+}
+
+::-webkit-scrollbar-thumb {
+  background: #444;
+  border-radius: 4px;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: #555;
+}
