+++ src/renderer/index.html
@@ -4,7 +4,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Image Gallery Manager</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="base.css">
+    <link rel="stylesheet" href="gallery.css">
+    <link rel="stylesheet" href="fullscreen.css">
+    <link rel="stylesheet" href="loading.css">
+    <link rel="stylesheet" href="archive-ui.css">
 </head>
 <body>
     <div id="app">
