--- src/renderer/styles.css
@@ -1,428 +1,0 @@
-/* Reset and base styles */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family:
-    -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;
-  background-color: #1a1a1a;
-  color: #ffffff;
-  overflow: hidden;
-  height: 100vh;
-}
-
-/* App container */
-#app {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-/* Gallery container */
-#gallery-container {
-  flex: 1;
-  position: relative;
-  overflow: hidden;
-}
-
-/* Drop zone */
-#drop-zone {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: linear-gradient(135deg, #2a2a2a 0%, #1a1a1a 100%);
-  border: 2px dashed #444;
-  transition:
-    background-color 0.3s ease,
-    border-color 0.3s ease;
-}
-
-#drop-zone.drag-over {
-  background: linear-gradient(135deg, #3a3a3a 0%, #2a2a2a 100%);
-  border-color: #666;
-}
-
-.drop-zone-content {
-  text-align: center;
-  max-width: 400px;
-}
-
-.drop-icon {
-  font-size: 4rem;
-  margin-bottom: 1rem;
-  opacity: 0.7;
-}
-
-.drop-zone-content h2 {
-  font-size: 1.5rem;
-  margin-bottom: 0.5rem;
-  font-weight: 300;
-}
-
-.drop-zone-content p {
-  color: #ccc;
-  margin-bottom: 2rem;
-  font-size: 0.9rem;
-}
-
-#file-select-btn {
-  background: #007acc;
-  color: white;
-  border: none;
-  padding: 12px 24px;
-  border-radius: 6px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-
-#file-select-btn:hover {
-  background: #005aa3;
-}
-
-/* Gallery grid */
-#gallery-grid {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow-y: auto;
-  padding: 2px;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 2px;
-  align-items: start;
-  overflow-anchor: auto;
-  contain: layout style;
-}
-
-/* Additional performance optimizations */
-
-.gallery-image {
-  width: 100%;
-  height: auto;
-  display: block;
-  background: #1a1a1a;
-  cursor: pointer;
-  contain: layout style paint;
-  /* Disable dragging */
-  -webkit-user-drag: none;
-  -webkit-user-select: none;
-  user-select: none;
-  /* Performance optimizations */
-  -webkit-backface-visibility: hidden;
-  backface-visibility: hidden;
-}
-
-/* No hover effects for better performance */
-
-.image-error {
-  width: 100%;
-  min-height: 200px; /* Give error placeholders a minimum height */
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background: #2a1a1a;
-  color: #ff6b6b;
-  font-size: 0.9rem;
-  text-align: center;
-  padding: 1rem;
-  border-radius: 2px;
-  border: 1px solid #333;
-  cursor: pointer;
-  contain: layout style paint;
-}
-
-.image-error-icon {
-  font-size: 2rem;
-  margin-bottom: 0.5rem;
-}
-
-/* Fullscreen overlay */
-#fullscreen-overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.95);
-  z-index: 1000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-#fullscreen-overlay.hidden {
-  display: none;
-}
-
-#fullscreen-image {
-  width: 100vw;
-  height: 100vh;
-  object-fit: contain;
-  background: #000;
-}
-
-#close-fullscreen {
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background: rgba(255, 255, 255, 0.2);
-  color: white;
-  border: none;
-  font-size: 18px;
-  cursor: pointer;
-  z-index: 20; /* above .nav-btn (z-index:10) */
-}
-
-.nav-btn {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background: rgba(255, 255, 255, 0.2);
-  color: white;
-  border: none;
-  padding: 20px 16px;
-  border-radius: 50%;
-  cursor: pointer;
-  font-size: 2rem;
-  transition: all 0.2s ease;
-  z-index: 10;
-}
-
-.nav-btn:hover {
-  background: rgba(255, 255, 255, 0.3);
-  transform: translateY(-50%) scale(1.1);
-}
-
-.prev-btn {
-  left: 20px;
-}
-
-.next-btn {
-  right: 20px;
-}
-
-.nav-btn:disabled {
-  opacity: 0.3;
-  cursor: not-allowed;
-}
-
-/* Loading indicator */
-#loading-indicator {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-  z-index: 2000;
-  background: rgba(0, 0, 0, 0.8);
-  padding: 20px;
-  border-radius: 10px;
-  color: white;
-  min-width: 250px;
-}
-
-/* Progress bar */
-#progress-bar {
-  width: 200px;
-  height: 8px;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 4px;
-  margin: 10px auto;
-  overflow: hidden;
-}
-
-#progress-fill {
-  height: 100%;
-  background: #007acc;
-  width: 0%;
-  transition: width 0.3s ease;
-}
-
-#progress-text {
-  font-size: 0.9rem;
-  color: #ccc;
-}
-
-.spinner {
-  width: 40px;
-  height: 40px;
-  border: 4px solid #333;
-  border-top: 4px solid #007acc;
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-  margin: 0 auto 1rem;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-/* Respect reduced motion preferences */
-@media (prefers-reduced-motion: reduce) {
-  .spinner {
-    animation: none;
-  }
-
-  html {
-    scroll-behavior: auto;
-  }
-
-  .nav-btn {
-    transition: none;
-  }
-
-  #progress-fill {
-    transition: none;
-  }
-
-  #drop-zone {
-    transition: none;
-  }
-
-  #file-select-btn {
-    transition: none;
-  }
-
-  .load-archive-btn {
-    transition: none;
-  }
-}
-
-/* Archive management styles */
-#archive-select-btn {
-  background: #28a745;
-  color: white;
-  border: none;
-  padding: 10px 15px;
-  border-radius: 5px;
-  cursor: pointer;
-  margin-top: 10px;
-  font-size: 14px;
-}
-
-#archive-select-btn:hover {
-  background: #218838;
-}
-
-#processed-archives-section {
-  margin-top: 20px;
-  padding: 15px;
-  background: rgba(255, 255, 255, 0.05);
-  border-radius: 8px;
-  border: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-#processed-archives-section h3 {
-  margin: 0 0 10px 0;
-  color: #fff;
-  font-size: 16px;
-}
-
-#processed-archives-list {
-  max-height: 300px;
-  overflow-y: auto;
-}
-
-.processed-archive-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 8px 12px;
-  margin: 5px 0;
-  background: rgba(255, 255, 255, 0.03);
-  border-radius: 4px;
-  border: 1px solid rgba(255, 255, 255, 0.05);
-}
-
-.archive-info {
-  flex: 1;
-}
-
-.archive-info strong {
-  display: block;
-  color: white;
-  font-size: 14px;
-}
-
-.archive-meta {
-  color: #ccc;
-  font-size: 12px;
-  margin-top: 2px;
-}
-
-.load-archive-btn {
-  background: #007acc;
-  color: white;
-  border: none;
-  padding: 6px 12px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 12px;
-  transition: background 0.2s;
-}
-
-.load-archive-btn:hover {
-  background: #005a9e;
-}
-
-/* Utility classes */
-.hidden {
-  display: none !important;
-}
-
-/* Fixed 5-column layout for now */
-
-/* Scrollbar styling */
-::-webkit-scrollbar {
-  width: 8px;
-}
-
-::-webkit-scrollbar-track {
-  background: #1a1a1a;
-}
-
-::-webkit-scrollbar-thumb {
-  background: #444;
-  border-radius: 4px;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background: #555;
-}
-
-/* Performance optimizations for smooth scrolling */
-* {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-html {
-  scroll-behavior: smooth;
-}
-
-/* Keyboard focus visibility */
-#close-fullscreen:focus-visible,
-.nav-btn:focus-visible,
-#file-select-btn:focus-visible,
-#archive-select-btn:focus-visible,
-.load-archive-btn:focus-visible {
-  outline: 2px solid #fff;
-  outline-offset: 3px;
-}
